analise_de_dados/requirements.txt
pandas==2.2.2
matplotlib==3.9.4
seaborn==0.13.2
jinja2==3.1.2
langchain==0.3.24
langchain-groq==0.3.8
langchain-experimental==0.3.4
python-dotenv==1.0.0